+++ Src/Services/Eshop.Orders/Models/UpdateCartItemDto.cs
@@ -8,7 +8,7 @@
 
         public int? Quantity { get; set; }
 
-        public double? FullPrice { get; set; }
+        public decimal? FullPrice { get; set; }
 
     }
 }
